Add support for multiple inheritance
Sadly, g++ has a bug that does not allow Introspective classes to
inherit from other Introspective classes. I already filed a report for
the bug, and it seems that the bug causes otherwise legal C++ to not
compile; we'll have to wait and see how g++ reacts.
g++ supports multiple inheritance from non-Introspective classes though,
so that is a good sign.